refactor(files): rename pages.rs → pages_admin.rs
